Refactor all kind of code, creating test-questions [2-17] Main audition module is ready
