+++ exalearn-ui/src/app/components/question/question.component.ts
@@ -1,11 +1,17 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { Question } from '../../interfaces/interfaces';
+import { questions } from '../../test-data/test-questions';
 
 @Component({
 	selector: 'app-question',
 	templateUrl: './question.component.html',
 	styleUrls: ['./question.component.scss']
 })
-export class QuestionComponent {
+export class QuestionComponent implements OnInit {
 	@Input() question: Question;
+	testQuestions: Question[] = [];
+
+	ngOnInit() {
+		this.testQuestions = questions;
+	}
 }
