+++ exalearn-ui/src/app/components/test-grammar/test-grammar.component.html
@@ -9,7 +9,7 @@
 			<mat-icon>arrow_left</mat-icon>
 		</button>
 		<app-test-nav-btn
-			*ngFor="let question of questions; index as i;"
+			*ngFor="let question of testQuestions; index as i;"
 			[question]="question"
 			[index]="i"
 			[currentIndex]="currentIndex"
@@ -20,12 +20,12 @@
 			mat-mini-fab
 			color="white"
 			class="nav-arrow-btn"
-			(click)="currentIndex < questions.length - 1 ? currentIndex = currentIndex + 1: currentIndex"
+			(click)="currentIndex < testQuestions.length - 1 ? currentIndex = currentIndex + 1: currentIndex"
 		>
 			<mat-icon>arrow_right</mat-icon>
 		</button>
 		<div>
-			<app-question [question]="questions[currentIndex]"></app-question>
+			<app-question [question]="testQuestions[currentIndex]"></app-question>
 		</div>
 	</div>
 </div>
