+++ exalearn-ui/src/app/enums/enums.ts
